alteração barcode para long

+++ APITCC/Models/ProductDetailRequest.cs
@@ -11,7 +11,7 @@ namespace APITCC.Models
         public string Description { get; set; }
         public string Price { get; set; }
         public DateTime? ExpirationDate { get; set; }
-        public int BarCode { get; set; }
+        public long BarCode { get; set; }
         public int StockQuantity { get; set; }
     }
 }
